incusd/storage/drivers: Handle 4k sector sizes
The way we run QEMU, we always provide 512 byte alignment to the VM, but
if the underlying physical storage (on LVM typically) is 4k, we need to
make sure that our call to sgdisk doesn't cause it to rewrite the
partition table for a 4k alignment.
